Move resources, database connections out of rest.py
